moved View to separated file
